nested loop

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,17 +6,17 @@
 */
 void print_fibonacci(void)
 {
-unsigned long term1 = 1;
-unsigned long term2 = 2;
-unsigned long nextTerm;
+unsigned long long term1 = 1;
+unsigned long long term2 = 2;
+unsigned long long nextTerm;
 unsigned int count;
 
-printf("%lu, %lu", term1, term2);
+printf("%llu, %llu", term1, term2);
 
 for (count = 3; count <= 98; count++)
 {
 	nextTerm = term1 + term2;
-	printf(", %lu", nextTerm);
+	printf(", %llu", nextTerm);
 
 	term1 = term2;
 	term2 = nextTerm;
@@ -35,5 +35,4 @@ int main(void)
 print_fibonacci();
 return (0);
 }
-
